+++ engine/webnodehonojs/src/webserver.js
@@ -46,7 +46,6 @@ module.exports = async (...args) => {
     const { dayjs, fs, path } = sys;
     const { existsSync, readFileSync } = fs;
     const { join } = path;
-    const { logpath } = cosetting;
 
     try {
       const cnttype = {
@@ -55,6 +54,7 @@ module.exports = async (...args) => {
       let lib = {};
       // let app = new OpenAPIHono();
       let app = new Hono();
+      let sessionval;
 
       /**
        * Web server establish and session log
@@ -66,17 +66,14 @@ module.exports = async (...args) => {
       const establish = (...args) => {
         let [setting] = args;
         try {
-          let {
-            logpath,
-            webnodejs: { parser, session, helmet },
-            general,
-          } = setting;
-          let { savestore, store, verbose, ...setsession } = session;
+          let { logpath, general } = setting;
+          let { savestore, store, verbose, ...setsession } =
+            setting[setting.args.engine].session;
 
           //set up our express application
           app.use(cors());
           app.use(secureHeaders());
-          
+
           // Compress all route
           // app.use(compress());
           if (savestore) {
@@ -120,8 +117,7 @@ module.exports = async (...args) => {
             },
           });
 
-          // Session in the middleware
-          app.use("*", sessionMiddleware(setsession));
+          sessionval = setsession;
 
           // 添加 Pino HTTP 中间件
           app.use("*", async (...args) => {
@@ -321,6 +317,9 @@ module.exports = async (...args) => {
               app,
             }),
           ]);
+
+          // Session in the middleware
+          app.use("*", sessionMiddleware(sessionval));
           app.use(
             bodyLimit({
               maxSize: setting.webbunjs.parser.maxSize,